[WIP] GenerateSAMLAssertionWithVerify init implementation
The SAMLService works, but in my effort to consolidate code a lot
of other things are probably broken including login.go. There's a
bit to clean up here, but this is the first step to consolidating
verifyFactor in device.go.

Note: we should create integration tests for login off of master
and ensure they still work here.