proxy server for saml assertion
This is a POC for proxying SAML Assertion requests via a
trusted intemediary server. The proxy configures a credentialed
onelogin client. The downside with this method is that a user's
password needs to be unpacked and retransmitted from the proxy,
which means security is paramount.

There other ways to skin this cat.