Feature/saml proxy server (#2)
* authentication integration tests

These integration tests include verify a complete login
authentication for the LoginService using the following methods:

* Authenticate
* AuthenticateWithVerify
* AuthenticateWithPushVerify (without verification)

Additionally, this adds a Makefile for linting, formatting and
running tests.

* [BREAKING] API change to remove thread-unsafe auth from LoginService

Without the breaking change to the login API, a user can be allowed
to potentially be authenticated as another user if the client is
used in a multi-threaded context. This changes the API to
explicitly pass back the auth information from PushVerify.

This also rename Devices to Device struct.

* proxy server for saml assertion

This is a POC for proxying SAML Assertion requests via a
trusted intemediary server. The proxy configures a credentialed
onelogin client. The downside with this method is that a user's
password needs to be unpacked and retransmitted from the proxy,
which means security is paramount.

There other ways to skin this cat.